+++ petri/src/vm/hyperv/powershell.rs
@@ -4,8 +4,9 @@
 //! Wrappers for Hyper-V Powershell Cmdlets
 
 use self::ps::PowerShellBuilder;
-use super::CommandError;
 use anyhow::Context;
+use cmd_builder::CommandError;
+use cmd_builder::ps;
 use core::str;
 use guid::Guid;
 use jiff::Timestamp;
@@ -731,331 +732,4 @@ pub fn run_remove_vm_scsi_controller(vmid: &Guid, controller_number: u32) -> any
         .output(true)
         .map(|_| ())
         .context("remove_vm_scsi_controller")
-}
-
-mod ps {
-    use crate::hyperv::CommandError;
-    use std::ffi::OsStr;
-    use std::ffi::OsString;
-    use std::path::Path;
-    use std::path::PathBuf;
-    use std::process::Command;
-    use std::process::Stdio;
-
-    /// A PowerShell script builder
-    pub struct PowerShellBuilder(Command);
-
-    impl PowerShellBuilder {
-        /// Create a new PowerShell command
-        pub fn new() -> Self {
-            PowerShellCmdletBuilder(Command::new("powershell.exe"))
-                .flag("NoProfile")
-                .finish()
-        }
-
-        /// Start a new Cmdlet
-        pub fn cmdlet<S: AsRef<str>>(self, cmdlet: S) -> PowerShellCmdletBuilder {
-            PowerShellCmdletBuilder(self.0).positional(RawVal::new(cmdlet.as_ref()))
-        }
-
-        /// Assign the output of the cmdlet to a variable
-        pub fn cmdlet_to_var<S: AsRef<str>>(
-            self,
-            cmdlet: S,
-            varname: &Variable,
-        ) -> PowerShellCmdletBuilder {
-            PowerShellCmdletBuilder(self.0)
-                .positional(varname)
-                .positional(RawVal::new("="))
-                .finish()
-                .cmdlet(cmdlet)
-        }
-
-        /// Run the PowerShell script and return the output
-        pub fn output(mut self, log_stdout: bool) -> Result<String, CommandError> {
-            self.0.stderr(Stdio::piped()).stdin(Stdio::null());
-
-            let ps_cmd = self.cmd();
-            tracing::debug!(ps_cmd, "executing powershell command");
-
-            let start = jiff::Timestamp::now();
-            let output = self.0.output()?;
-            let time_elapsed = jiff::Timestamp::now() - start;
-
-            let ps_stdout = (log_stdout || !output.status.success())
-                .then(|| String::from_utf8_lossy(&output.stdout).to_string());
-            let ps_stderr = String::from_utf8_lossy(&output.stderr).to_string();
-            tracing::debug!(
-                ps_cmd,
-                ps_stdout,
-                ps_stderr,
-                "powershell command exited in {:.3}s with status {}",
-                time_elapsed.total(jiff::Unit::Second).unwrap_or(-1.0),
-                output.status
-            );
-
-            if !output.status.success() {
-                return Err(CommandError::Command(output.status, ps_stderr));
-            }
-
-            Ok(String::from_utf8(output.stdout)?.trim().to_owned())
-        }
-
-        /// Get the command to be run
-        pub fn cmd(&self) -> String {
-            format!(
-                "{} {}",
-                self.0.get_program().to_string_lossy(),
-                self.0
-                    .get_args()
-                    .collect::<Vec<_>>()
-                    .join(OsStr::new(" "))
-                    .to_string_lossy()
-            )
-        }
-    }
-
-    /// A PowerShell Cmdlet builder
-    pub struct PowerShellCmdletBuilder(Command);
-
-    impl PowerShellCmdletBuilder {
-        /// Add a flag to the cmdlet
-        pub fn flag<S: AsRef<OsStr>>(mut self, flag: S) -> Self {
-            let mut arg = OsString::from("-");
-            arg.push(flag);
-            self.0.arg(arg);
-            self
-        }
-
-        /// Optionally add a flag to the cmdlet
-        pub fn flag_opt<S: AsRef<OsStr>>(self, flag: Option<S>) -> Self {
-            if let Some(flag) = flag {
-                self.flag(flag)
-            } else {
-                self
-            }
-        }
-
-        /// Add a positional argument to the cmdlet
-        pub fn positional<S: AsVal>(mut self, positional: S) -> Self {
-            self.0.arg(positional.as_val());
-            self
-        }
-
-        /// Optionally add a positional argument to the cmdlet
-        #[expect(dead_code)]
-        pub fn positional_opt<S: AsVal>(self, positional: Option<S>) -> Self {
-            if let Some(positional) = positional {
-                self.positional(positional)
-            } else {
-                self
-            }
-        }
-
-        /// Add a named argument to the cmdlet
-        pub fn arg<S: AsRef<OsStr>, T: AsVal>(self, name: S, value: T) -> Self {
-            self.flag(name).positional(value)
-        }
-
-        /// Optionally add a named argument to the cmdlet
-        pub fn arg_opt<S: AsRef<OsStr>, T: AsVal>(self, name: S, value: Option<T>) -> Self {
-            if let Some(value) = value {
-                self.arg(name, value)
-            } else {
-                self
-            }
-        }
-
-        /// Finish the cmdlet
-        pub fn finish(self) -> PowerShellBuilder {
-            PowerShellBuilder(self.0)
-        }
-
-        /// Finish the cmdlet with a pipeline operator
-        pub fn pipeline(mut self) -> PowerShellBuilder {
-            self.0.arg("|");
-            self.finish()
-        }
-
-        /// Finish the cmdlet with a semicolon
-        pub fn next(mut self) -> PowerShellBuilder {
-            self.0.arg(";");
-            self.finish()
-        }
-    }
-
-    pub struct Value(OsString);
-
-    impl Value {
-        pub fn new(val: impl AsVal) -> Self {
-            Self(val.as_val().as_ref().to_owned())
-        }
-    }
-
-    impl AsVal for Value {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-            &self.0
-        }
-    }
-
-    pub trait AsVal {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr>;
-    }
-
-    impl<T: AsVal + ?Sized> AsVal for &T {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-            (*self).as_val()
-        }
-    }
-
-    fn quote_str(s: &OsStr) -> OsString {
-        let mut quoted = OsString::new();
-        quoted.push("\"");
-        // TODO: escape this properly.
-        quoted.push(s);
-        quoted.push("\"");
-        quoted
-    }
-
-    macro_rules! str {
-        ($($ty:ty),* $(,)?) => {
-            $(
-                impl AsVal for $ty {
-                    fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-                        quote_str(self.as_ref())
-                    }
-                }
-            )*
-        }
-    }
-
-    str!(&str, String, Path, PathBuf);
-
-    macro_rules! disp_str {
-        ($($ty:ty),* $(,)?) => {
-            $(
-                impl AsVal for $ty {
-                    fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-                        quote_str(self.to_string().as_ref())
-                    }
-                }
-            )*
-        }
-    }
-
-    disp_str!(jiff::Timestamp, guid::Guid);
-
-    impl AsVal for bool {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-            if *self { "$true" } else { "$false" }
-        }
-    }
-
-    macro_rules! disp {
-        ($($ty:ty),* $(,)?) => {
-            $(
-                impl AsVal for $ty {
-                    fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-                        self.to_string()
-                    }
-                }
-            )*
-        }
-    }
-
-    disp!(u8, u16, u32, u64, i8, i16, i32, i64, f32, f64);
-
-    pub struct RawVal<T>(T);
-
-    impl<T: AsRef<OsStr>> RawVal<T> {
-        pub fn new(arg: T) -> Self {
-            Self(arg)
-        }
-    }
-
-    impl<T: AsRef<OsStr>> AsVal for RawVal<T> {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-            &self.0
-        }
-    }
-
-    pub struct Variable(String);
-
-    impl Variable {
-        pub fn new(name: impl AsRef<str>) -> Self {
-            Self(format!("${}", name.as_ref()))
-        }
-    }
-
-    impl AsVal for Variable {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-            &self.0
-        }
-    }
-
-    pub struct Array(OsString);
-
-    impl Array {
-        pub fn new<T: AsVal>(v: impl IntoIterator<Item = T>) -> Self {
-            let mut args = OsString::new();
-            args.push("@(");
-            let mut first = true;
-            for arg in v {
-                if !first {
-                    args.push("; ");
-                }
-                args.push(arg.as_val());
-                first = false;
-            }
-            args.push(")");
-            Self(args)
-        }
-    }
-
-    impl AsVal for Array {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-            &self.0
-        }
-    }
-
-    pub struct HashTable<K, V>(Vec<(K, V)>);
-
-    impl<K: AsRef<str>, V: AsVal> HashTable<K, V> {
-        pub fn new(v: impl IntoIterator<Item = (K, V)>) -> Self {
-            Self(v.into_iter().collect())
-        }
-    }
-
-    impl<K: AsRef<str>, V: AsVal> AsVal for HashTable<K, V> {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-            let mut args = OsString::new();
-            args.push("@{");
-            let mut first = true;
-            for (k, v) in &self.0 {
-                if !first {
-                    args.push("; ");
-                }
-                args.push(k.as_ref());
-                args.push("=");
-                args.push(v.as_val());
-                first = false;
-            }
-            args.push("}");
-            args
-        }
-    }
-
-    pub struct Script(String);
-
-    impl Script {
-        pub fn new(script: impl AsRef<str>) -> Self {
-            Self(format!("{{ {} }}", script.as_ref()))
-        }
-    }
-
-    impl AsVal for Script {
-        fn as_val(&self) -> impl '_ + AsRef<OsStr> {
-            &self.0
-        }
-    }
 }
