+++ support/cmd_builder/Cargo.toml
@@ -1,0 +1,15 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+[package]
+name = "cmd_builder"
+edition.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+jiff = { workspace = true, features = ["serde"] }
+thiserror.workspace = true
+tracing.workspace = true
+
+[lints]
+workspace = true
